Updates with new algorithm for task 2 (scoring system)
Hey all,

I've placed the updated java file + the findings for task 2 in the root
directory for your convenience.

The algorithm used is a scoring system that with the following steps:
1. Tries to match the candidate mac addresses with the hardcoded ones
2. Assigns a score for each matched MAC address based on the similarity
of the RSSI values between that and the hardcoded ones. (Note: More
similar scores will get the candidate more points, i.e. if candidate's
RSSI value for MAC address A = 85 whereas hardcoded RSSI value = 90, we
would give it a higher score as compared to if candidate's RSSI value
for MAC address A = 70)
3. Compare resulting scores and extract the highest scoring point as the
localized location.

Think the above algorithm is quite accurate at the moment since all 13
MAC addresses of the hardcoded version matches our current best
candidate && more importantly, the differences between the RSSI values
are <=6 each! :)

Do feel to refine the algorithm/discuss on the group chat :D Thanks!!